sched: fixed a 32 bits only misplanification for tasks planned in > 1 month
When scheduling a task in more than 2^31 milliseconds (about 1 month and a
few days) the QTimer was not configured (because the 64 bits to 32 bits
conversion resulted in a negative number) and therefore only be started
by periodic check, therefore 0 to 60 seconds after the exact planned
date.

Now the QTimer is set to its maximum (2^31 ms) and will fire and be recomputed
every 2^31 ms until the exact planned date.

The bug did not affect 64 bits targets.