sched: introducing "each" cluster balancing method
This is not an actual balancing method but rather a mean to declare a
task as beeing executed on every member host of a cluster.

When requesting a task targeted at such a cluster, instead of queuing
one task request, the scheduler queues as many request as hosts in the
cluster, each of it targeted on one host.